refactor(cli): move NpmOptions and PackageJson in their own files
