+++ lib/cli/src/NpmOptions.ts
@@ -1,0 +1,5 @@
+export type NpmOptions = {
+  useYarn: boolean;
+  skipInstall?: boolean;
+  installAsDevDependencies?: boolean;
+};
