+++ lib/cli/src/PackageJson.ts
@@ -1,0 +1,5 @@
+export type PackageJson = {
+  dependencies?: Record<string, string>;
+  devDependencies?: Record<string, string>;
+  peerDependencies?: Record<string, string>;
+};
